refactor: Refactor component names
